+++ NEWS.md
@@ -19,6 +19,8 @@ Version 0.18b0 - dev
   divide polylines and anything that can be approximated or flattened into 
   vertices
 - NEW: approximation tool for parametrized curves: `ezdxf.math.ApproxParamT()`
+- NEW: `BoundingBox.intersection(other)`, returns the 3D bbox of the intersection cube
+- NEW: `BoundingBox2d.intersection(other)`, returns the 2D bbox of the intersection rectangle
 - NEW: `ezdxf.gfxattribs.GfxAttribs()` class, [docs](https://ezdxf.mozman.at/docs/tools/gfxattribs.html)
 - NEW: `TextEntityAlignment` enum replaces the string based alignment definition
 - NEW: method `Text.get_placement()`, replaces `get_pos()`